workflows/tuning/requirements.txt
parameter-sherpa==1.0.6
pyyaml==5.1.1
kubernetes==10.0.1
